+++ ObjectPool/Specialized/IMemoryStreamPool.cs
@@ -1,4 +1,4 @@
-﻿// File name: MemoryStreamPool.cs
+﻿// File name: IMemoryStreamPool.cs
 //
 // Author(s): Alessio Parma <alessio.parma@gmail.com>
 //
@@ -21,73 +21,26 @@
 // DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT
 // OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using CodeProject.ObjectPool.Core;
 using System.IO;
 
 namespace CodeProject.ObjectPool.Specialized
 {
     /// <summary>
-    ///   An <see cref="IObjectPool{PooledMemoryStream}"/> ready to be used.
-    ///   <see cref="MemoryStream"/> management can be further configured using the
-    ///   <see cref="MinimumMemoryStreamCapacity"/> and <see cref="MaximumMemoryStreamCapacity"/> properties.
+    ///   An object pool specialized in <see cref="MemoryStream"/> management.
     /// </summary>
-    public sealed class MemoryStreamPool : ObjectPool<PooledMemoryStream>, IMemoryStreamPool
+    public interface IMemoryStreamPool : IObjectPool<PooledMemoryStream>
     {
-        private int _minimumItemCapacity = SpecializedPoolConstants.DefaultMinimumMemoryStreamCapacity;
-        private int _maximumItemCapacity = SpecializedPoolConstants.DefaultMaximumMemoryStreamCapacity;
-
-        /// <summary>
-        ///   Thread-safe pool instance.
-        /// </summary>
-        public static IMemoryStreamPool Instance { get; } = new MemoryStreamPool();
-
-        /// <summary>
-        ///   Builds the specialized pool.
-        /// </summary>
-        public MemoryStreamPool()
-            : base(ObjectPoolConstants.DefaultPoolMinimumSize, ObjectPoolConstants.DefaultPoolMaximumSize, null, false)
-        {
-            FactoryMethod = () => new PooledMemoryStream(MinimumMemoryStreamCapacity);
-            AdjustPoolSizeToBounds(AdjustMode.Minimum | AdjustMode.Maximum);
-        }
-
         /// <summary>
         ///   Minimum capacity a <see cref="MemoryStream"/> should have when created and this is the
         ///   minimum capacity of all streams stored in the pool. Defaults to <see cref="SpecializedPoolConstants.DefaultMinimumMemoryStreamCapacity"/>.
         /// </summary>
-        public int MinimumMemoryStreamCapacity
-        {
-            get { return _minimumItemCapacity; }
-            set
-            {
-                var oldValue = _minimumItemCapacity;
-                _minimumItemCapacity = value;
-                if (oldValue < value)
-                {
-                    Clear();
-                }
-            }
-        }
+        int MinimumMemoryStreamCapacity { get; set; }
 
         /// <summary>
         ///   Maximum capacity a <see cref="MemoryStream"/> might have in order to be able to return
         ///   to pool. Defaults to <see cref="SpecializedPoolConstants.DefaultMaximumMemoryStreamCapacity"/>.
         /// </summary>
-        public int MaximumMemoryStreamCapacity
-        {
-            get { return _maximumItemCapacity; }
-            set
-            {
-                var oldValue = _maximumItemCapacity;
-                _maximumItemCapacity = value;
-                if (oldValue > value)
-                {
-                    Clear();
-                }
-            }
-        }
-
-#pragma warning disable CC0022 // Should dispose object
+        int MaximumMemoryStreamCapacity { get; set; }
 
         /// <summary>
         ///   Returns a pooled memory stream using given byte array as stream buffer. Once the object
@@ -99,11 +52,6 @@ namespace CodeProject.ObjectPool.Specialized
         ///   When you pass a buffer to this method, you lose the ownership of the buffer, since it
         ///   might be claimed by the pool itself.
         /// </remarks>
-        public PooledMemoryStream GetObject(byte[] buffer) => new PooledMemoryStream(buffer)
-        {
-            Handle = this
-        };
-
-#pragma warning restore CC0022 // Should dispose object
+        PooledMemoryStream GetObject(byte[] buffer);
     }
 }